# modernized/shared/openapi.yaml
openapi: 3.0.3
info:
  title: CardDemo API
  description: >
    REST API contract for the modernized CardDemo credit card management application.
    Forward-engineered from the legacy CICS/BMS transactions per docs/spec/**.
    This file is the source of truth consumed by @carddemo/backend (implementation)
    and @carddemo/frontend (client).
  version: 0.1.0
servers:
  - url: /api/v1
...
paths:
...
  /cards:
    get:
      tags: [cards]
      summary: List cards, filterable by account, paginated (legacy COCRDLIC)
      operationId: listCards
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
            pattern: '^\d{11}$'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paginated card list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Card'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cards/{cardNumber}:
    parameters:
      - $ref: '#/components/parameters/CardNumber'
    get:
      tags: [cards]
      summary: View card details (legacy COCRDSLC)
      operationId: getCard
      responses:
        '200':
          description: Card details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [cards]
      summary: Update card name, expiry, or status (legacy COCRDUPC)
      operationId: updateCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        '200':
          description: Updated card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CardNumber:
      name: cardNumber
      in: path
      required: true
      schema:
        type: string
        pattern: '^\d{16}$'
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
          example: Bad Request
        message:
          type: string
        details:
          type: array
          description: Field-level validation issues, when applicable
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
    Page:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
...
    IsoDate:
      type: string
      format: date
      example: '2024-01-15'
...
    Card:
      type: object
      required: [cardNumber, accountId, embossedName, expiryDate, activeStatus]
      properties:
        cardNumber:
          type: string
          pattern: '^\d{16}$'
        accountId:
          type: string
          pattern: '^\d{11}$'
        embossedName:
          type: string
          maxLength: 50
        expiryDate:
          $ref: '#/components/schemas/IsoDate'
        activeStatus:
          type: boolean
    CardUpdateRequest:
      type: object
      properties:
        embossedName:
          type: string
          maxLength: 50
        expiryDate:
          $ref: '#/components/schemas/IsoDate'
        activeStatus:
          type: boolean
...
security:
  - bearerAuth: []